refactor: interface, enum 위치 변경
