+++ src/components/atoms/Select/Select.tsx
@@ -11,7 +11,6 @@ export interface SelectProps {
   defaultName?: string;
 }
 
-// TODO: 상행선/하행선 라벨로 달아서 표기하기
 const Select = ({ defaultName, options, onChange, selectValue, ...props }: SelectProps) => (
   <Container onChange={onChange} value={selectValue} {...props}>
     {defaultName && <option hidden>{defaultName}</option>}
