+++ src/constants/index.ts
@@ -1,5 +1,6 @@
 import { GUIDE_MESSAGE } from './message';
 import ROUTE from './route';
 import HOST from './host';
+import LineColor from './color';
 
-export { ROUTE, GUIDE_MESSAGE, HOST };
+export { ROUTE, GUIDE_MESSAGE, HOST, LineColor };
